Briac you are a legend!

+++ Dependencies/UI/include/UIManager.h
@@ -33,8 +33,8 @@ public:
 	void mousePositionInput(double x, double y);
 	void mouseButtonInput(int button, int action);
 	void render();
-	int getWidth() { return width; }
-	int getHeight() { return height; }
+	float getWidth() { return width; }
+	float getHeight() { return height; }
     ~UIManager();
 private:
     Mat4 ortho;
